Convert orchestrator integration tests to in-process harness
Closes #1060.

Every `crates/harn-cli/tests/orchestrator_*` test now runs the
orchestrator in-process via `OrchestratorHarness` (added in #1059) and
waits on the event log's broadcast channel, replacing subprocess
spawn-and-poll with deterministic event-driven waits.

Mechanical scope:
- `orchestrator_http/{a2a,admin,github,notion,slack}.rs`: 17 tests
  cut over to `start_harness` + `await_topic_event` + `harness.shutdown`
- `orchestrator_http/support.rs` rewritten as a shared in-process
  helper module (manifest fixtures, env-locked harness start, event
  log subscribe wrappers, signing helpers); `lock_orchestrator_tests`,
  `wait_for_topic_event`, `wait_for_readyz`, `wait_for_listener_url`,
  `state_snapshot`, `spawn_orchestrator`, `send_sigterm`, and the
  `OrchestratorProcess` wrapper are all gone
- `orchestrator_cli.rs`: 7 tests cut over (graceful drain, pump error
  recovery, inbox backpressure, drain-truncate-then-replay restart,
  legacy soft-migration); `worker_queue_drain` keeps using
  `harn_command` because it only exercises CLI surface, never the
  orchestrator daemon
- `orchestrator_http/observability.rs` and `orchestrator_inbox_dedupe`
  collapse to `#![cfg(any())]` placeholders: their assertions depend
  on process-global tracing-subscriber install or
  `std::process::exit(86)` crash hooks, both of which are subprocess-
  only. Move tracked under #1069 (slow E2E/smoke job)
- `orchestrator_cli::orchestrator_serve_starts_and_shuts_down_cleanly`
  and `restart_surfaces_stranded_envelopes_*` keep `#[ignore]`
  stubs pointing at #1069 for the same reason
- `tests/support/mod.rs` deleted; the surviving `support/mcp.rs` and
  `support/process.rs` (used by `mcp_server_cli` and
  `harn_serve_mcp_cli`, which test the MCP CLI surface and stay out of
  scope per the issue) are unchanged

Library-side plumbing required to expose the harness to integration
tests in `tests/`:
- `crates/harn-cli/src/lib.rs` added with the module declarations
  previously in `main.rs`; `main.rs` becomes a 3-line shim
  (`harn_cli::run()`)
- `commands::orchestrator::{harness,tls}` and `tests::common::{cwd_lock,
  env_lock,harn_state_lock}` promoted to `pub`; `env_guard::ScopedEnvVar`
  promoted to `pub`; `OrchestratorRole`/`TlsFiles`/`AdminReloadHandle`
  upgraded from `pub(crate)` to `pub` so they can appear in the
  harness's public surface
- `OrchestratorConfig::for_test` loses the `#[cfg(test)]` gate so
  integration tests can construct it
- `TlsFiles::new(cert, key)` constructor added so the TLS test no
  longer needs access to private fields

Acceptance criteria (verified):
- `grep -r "spawn_orchestrator\|wait_for_listener_url\|
  wait_for_topic_event\|wait_for_readyz" crates/harn-cli/tests/`
  returns zero hits
- 10/10 consecutive `cargo nextest run -p harn-cli --test
  orchestrator_http --test orchestrator_cli --test
  orchestrator_inbox_dedupe` runs pass at default thread count
- Identical results at `-j 1` and `-j 16`
- `orchestrator_http` suite drops from minutes (subprocess startup +
  25 ms-poll drain per test) to ~13 s end-to-end
- `cargo clippy --workspace --all-targets -- -D warnings` clean
- `cargo test -p harn-cli` green (the unrelated
  `commands::mcp::serve::tests::file_backed_prompts_list_render_and_notify_changes`
  flake reproduces on `main` and is out of scope here)
