+++ crates/harn-cli/src/env_guard.rs
@@ -1,16 +1,16 @@
-pub(crate) struct ScopedEnvVar {
+pub struct ScopedEnvVar {
     key: &'static str,
     previous: Option<String>,
 }
 
 impl ScopedEnvVar {
-    pub(crate) fn set(key: &'static str, value: &str) -> Self {
+    pub fn set(key: &'static str, value: &str) -> Self {
         let previous = std::env::var(key).ok();
         std::env::set_var(key, value);
         Self { key, previous }
     }
 
-    pub(crate) fn set_if_unset(key: &'static str, value: &str) -> Option<Self> {
+    pub fn set_if_unset(key: &'static str, value: &str) -> Option<Self> {
         if std::env::var(key).is_ok() {
             None
         } else {
