+++ crates/harn-cli/src/tests/common/env_lock.rs
@@ -5,6 +5,6 @@ use tokio::sync::Mutex;
 static ENV_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
 
 /// Serialize tests that mutate process-wide environment variables.
-pub(crate) fn lock_env() -> &'static Mutex<()> {
+pub fn lock_env() -> &'static Mutex<()> {
     ENV_LOCK.get_or_init(|| Mutex::new(()))
 }
