+++ crates/harn-cli/src/tests/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod common;
+pub mod common;
