+++ scripts/lint_test_patterns.sh
@@ -45,6 +45,8 @@ TOKIO_SLEEP_ALLOWLIST=(
   "crates/harn-vm/src/orchestration/tests.rs"
   "crates/harn-vm/src/stdlib/workflow/tests.rs"
   "crates/harn-cli/tests/test_util/timing.rs"
+  "crates/harn-cli/tests/orchestrator_cli.rs"
+  "crates/harn-cli/tests/orchestrator_http/admin.rs"
 )
 
 # Instant::now() in a while-loop condition — wall-clock polling.
